Created the api and started importing the info
Huge update, I finally was able to create the api
and import it into the main file. I'm super
excited now because it is the first api I've ever
created and it's working perfectly. I was able to
get the information from the api on this file now.
It doesn't look the best but it's a start. Will
work on this more later.